requirements.txt
streamlit
sounddevice
wavio
openai-whisper
python-Levenshtein